Double parsing issue

+++ BusinessLogic/Models/Offer.cs
@@ -6,10 +6,10 @@ namespace BusinessLogic.Models
     public class Offer
     {
         public string Title { get; set; }
-        public decimal Area { get; set; }
+        public double Area { get; set; }
         public AreaUnit AreaUnit { get; set; }
-        public decimal PricePerUnit { get; set; }
-        public decimal Price { get; set; }
+        public double PricePerUnit { get; set; }
+        public double Price { get; set; }
 
         public string Url { get; set; }
         public string Location { get; set; }
